Move relay code over to utils.py
